requirements.txt
pillow
opencv-contrib-python
web.py==0.40.dev0
torch 
torchvision
editdistance
xlsxwriter
